CodeDiffs

Compare different types of code and display it in the terminal. For cleaner results, syntax highlighting is separated from the difference calculation.

Supports:

native CPU assembly (output of @code_native, highlighted by InteractiveUtils.print_native)

LLVM IR (output of @code_llvm, highlighted by InteractiveUtils.print_llvm)

Typed Julia IR (output of @code_typed, highlighted through the Base.show method of Core.CodeInfo)

Julia AST (an Expr), highlighting is done with:

OhMyREPL.jl's Julia syntax highlighting in Markdown code blocks

(Julia ≥ v1.11) JuliaSyntaxHighlighting.jl

The @code_diff macro is the main entry point. If possible, the code type will be detected automatically, otherwise add e.g. type=:native for native assembly comparison:

julia> f1(a) = a + 1
f1 (generic function with 1 method)

julia> @code_diff type=:llvm debuginfo=:none color=false f1(Int64(1)) f1(Int8(1))
define i64 @f1(i64 signext %0) #0 {   ⟪╋⟫define i64 @f1(i8 signext %0) #0 {
top:                                   ┃ top:
                                       ┣⟫  %1 = sext i8 %0 to i64
  %1 = add i64 %0, 1                  ⟪╋⟫  %2 = add nsw i64 %1, 1
  ret i64 %1                          ⟪╋⟫  ret i64 %2
}                                      ┃ }

julia> f2(a) = a - 1
f2 (generic function with 1 method)

julia> @code_diff type=:llvm debuginfo=:none color=false f1(1) f2(1)
define i64 @f1(i64 signext %0) #0 {   ⟪╋⟫define i64 @f2(i64 signext %0) #0 {
top:                                   ┃ top:
  %1 = add i64 %0, 1                  ⟪╋⟫  %1 = add i64 %0, -1
  ret i64 %1                           ┃   ret i64 %1
}                                      ┃ }

Setting the environment variable "CODE_DIFFS_LINE_NUMBERS" to true will display line numbers on each side.

Main functions

CodeDiffs.CodeDiff — Type

CodeDiff(code₁, code₂)
CodeDiff(code₁, code₂, highlighted₁, highlighted₂)

A difference between code₁ and code₂.

code₁ and code₂ should have no highlighting. Only highlighted₁ and highlighted₂ should have syntax highlighting. When showing the differences, their formatting will be re-applied.

For cleaner differences, use replace_llvm_module_name on all codes.

Use optimize_line_changes! to improve the difference.

Fancy REPL output is done with side_by_side_diff.

CodeDiffs.compare_code_native — Function

compare_code_native(code₁, code₂; color=true)

Return a CodeDiff between code₁ and code₂. Codes are cleaned-up with replace_llvm_module_name beforehand.

If color == true, then both codes are highlighted using InteractiveUtils.print_native.

compare_code_native(
    f₁::Base.Callable, types₁::Type{<:Tuple},
    f₂::Base.Callable, types₂::Type{<:Tuple};
    color=true, kwargs...
)

Call InteractiveUtils.code_native(f₁, types₁) and InteractiveUtils.code_native(f₂, types₂) and return their CodeDiff. kwargs are passed to code_native.

compare_code_native(
    f::Base.Callable, types::Type{<:Tuple}, world₁, world₂;
    color=true, kwargs...
)

Similar to compare_code_native(f₁, types₁, f₂, types₂), but as a difference between f in world ages world₁ and world₂.

CodeDiffs.compare_code_llvm — Function

compare_code_llvm(code₁, code₂; color=true)

Return a CodeDiff between code₁ and code₂. Codes are cleaned-up with replace_llvm_module_name beforehand.

If color == true, then both codes are highlighted using InteractiveUtils.print_llvm.

compare_code_llvm(
    f₁::Base.Callable, types₁::Type{<:Tuple},
    f₂::Base.Callable, types₂::Type{<:Tuple};
    color=true, kwargs...
)

Call InteractiveUtils.code_llvm(f₁, types₁) and InteractiveUtils.code_llvm(f₂, types₂) and return their CodeDiff. kwargs are passed to code_llvm.

compare_code_llvm(
    f::Base.Callable, types::Type{<:Tuple}, world₁, world₂;
    color=true, kwargs...
)

Similar to compare_code_llvm(f₁, types₁, f₂, types₂), but as a difference between f in world ages world₁ and world₂.

CodeDiffs.compare_code_typed — Function

compare_code_typed(code_info₁::Pair, code_info₂::Pair; color=true)
compare_code_typed(code_info₁::Core.CodeInfo, code_info₂::Core.CodeInfo; color=true)

Return a CodeDiff between code_info₁ and code_info₂.

If color == true, then both codes are highlighted.

compare_code_typed(
    f₁::Base.Callable, types₁::Type{<:Tuple},
    f₂::Base.Callable, types₂::Type{<:Tuple};
    color=true, kwargs...
)

Call Base.code_typed(f₁, types₁) and Base.code_typed(f₂, types₂) and return their CodeDiff. kwargs are passed to code_typed.

Both function calls should only match a single method.

compare_code_typed(
    f::Base.Callable, types::Type{<:Tuple}, world₁, world₂;
    color=true, kwargs...
)

Similar to compare_code_typed(f₁, types₁, f₂, types₂), but as a difference between f in world ages world₁ and world₂.

CodeDiffs.compare_ast — Function

compare_ast(code₁::Expr, code₂::Expr; color=true, prettify=true, lines=false, alias=false)

A CodeDiff between code₁ and code₂, relying on the native display of Julia AST.

If prettify == true, then MacroTools.prettify(code; lines, alias) is used to cleanup the AST. lines == true will keep the LineNumberNodes and alias == true will replace mangled names (or gensyms) by dummy names.

color == true is special, as it places the stringified AST into a Markdown code block. See compare_ast(code₁::Markdown.MD, code₂::Markdown.MD).

compare_ast(code₁::Markdown.MD, code₂::Markdown.MD; color=true)

CodeDiff between Julia code string, in the form of Markdown code blocks.

Relies on the Markdown code highlighting from OhMyREPL.jl to colorize Julia code.

CodeDiffs.code_diff — Method

code_diff(::Val{type}, f₁, types₁, f₂, types₂; kwargs...)
code_diff(::Val{type}, code₁::AbstractString, code₂::AbstractString; kwargs...)
code_diff(args...; type=:native, kwargs...)

Dispatch to compare_code_native, compare_code_llvm, compare_code_typed or compare_ast depending on type.

CodeDiffs.code_diff — Method

code_diff(code₁::Markdown.MD, code₂::Markdown.MD; kwargs...)
code_diff(code₁::Expr, code₂::Expr; kwargs...)

Compare AST in code₁ and code₂. Expr are placed in Markdown code blocks.

CodeDiffs.@code_diff — Macro

@code_diff [type=:native] [option=value...] f₁(...) f₂(...)
@code_diff [type] [option=value...] a b

Compare the methods called by the f₁(...) and f₂(...) expressions, and return a CodeDiff. In the other form of @code_diff, a and b must be either variable names (Symbols) or quoted expressions (e.g. @code_diff :(1+2) :(2+3)): in this case the difference type might be inferred automatically.

options are passed as key-word arguments to code_diff and then to the compare_code_* function for the given code type.

Display functions

CodeDiffs.optimize_line_changes! — Function

optimize_line_changes!(diff::CodeDiff; dist=Levenshtein(), tol=0.7)

Merges consecutive line removals+additions into single line changes in diff, when they are within the tolerance of the normalized string distance.

This does not aim to produce an optimal CodeDiff, but simply improve its display.

CodeDiffs.replace_llvm_module_name — Function

replace_llvm_module_name(code::AbstractString)

Remove in code the trailing numbers in the LLVM module names, e.g. "julia_f_2007" => "f". This allows to remove false differences when comparing raw code, since each call to code_native (or code_llvm) triggers a new compilation using an unique LLVM module name, therefore each consecutive call is different even though the actual code does not change.

julia> f() = 1
f (generic function with 1 method)

julia> buf = IOBuffer();

julia> code_native(buf, f, Tuple{})  # Equivalent to `@code_native f()`

julia> code₁ = String(take!(buf));

julia> code_native(buf, f, Tuple{})

julia> code₂ = String(take!(buf));

julia> code₁ == code₂  # Different LLVM module names...
false

julia> replace_llvm_module_name(code₁) == replace_llvm_module_name(code₂)  # ...but same code
true

replace_llvm_module_name(code::AbstractString, function_name)

Replace only LLVM module names for function_name.

CodeDiffs.side_by_side_diff — Function

side_by_side_diff([io::IO,] diff::CodeDiff; tab_width=4, width=nothing, line_numbers=nothing)

Side by side display of a CodeDiff to io (defaults to stdout).

width defaults to the width of the terminal. It is 80 by default for non-terminal io.

tab_width is the number of spaces tabs are replaced with.

line_numbers=true will add line numbers on each side of the columns. It defaults to the environment variable "CODE_DIFFS_LINE_NUMBERS", which itself defaults to false.

Internals

CodeDiffs.LLVM_MODULE_NAME_REGEX — Constant

LLVM_MODULE_NAME_REGEX

Should match the LLVM module of any function which does not have any of '",;- or spaces in it.

It is 'get_function_name', in 'julia/src/codegen.cpp' which builds the function name for the LLVM module used to get the function code. The regex is built to match any output from that function. Since the 'globalUniqueGeneratedNames' counter (the number at the end of the module name) is incremented at each call to 'get_function_name', and since code_llvm or code_native forces a compilation, it should be guaranteed that the match with the highest number at the end is the name of our function in code.